+++ src/semble/index/index.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import os
 import subprocess
 import tempfile
+import warnings
 from collections import defaultdict
 from collections.abc import Sequence
 from pathlib import Path
@@ -11,13 +12,24 @@ import numpy as np
 import numpy.typing as npt
 from bm25s import BM25
 
-from semble.index.create import _apply_include_text_files, create_index_from_path
+from semble.index.create import create_index_from_path
 from semble.index.dense import SelectableBasicBackend, load_model
 from semble.search import DEFAULT_DOCS_DIVERSITY, _search_semantic, search
 from semble.stats import save_search_stats
 from semble.types import CallType, Chunk, ContentType, Encoder, IndexStats, SearchResult
 
 _GIT_CLONE_TIMEOUT = int(os.environ.get("SEMBLE_CLONE_TIMEOUT", 60))
+_INCLUDE_TEXT_FILES_DEPRECATION_MSG = (
+    "include_text_files is deprecated and will be removed in a future version. Use content=ContentType.ALL instead."
+)
+
+
+def _apply_include_text_files(content: ContentType, include_text_files: bool | None) -> ContentType:
+    """Apply the deprecated include_text_files override, emitting a DeprecationWarning."""
+    if include_text_files is None:
+        return content
+    warnings.warn(_INCLUDE_TEXT_FILES_DEPRECATION_MSG, DeprecationWarning, stacklevel=3)
+    return ContentType.ALL if include_text_files else ContentType.CODE
 
 
 class SembleIndex:
@@ -39,7 +51,7 @@ class SembleIndex:
         :param semantic_index: The semantic index.
         :param chunks: The found chunks.
         :param root: Root directory used to read file sizes for token-savings stats.
-        :param content: Content types used when indexing; controls the search pipeline.
+        :param content: Content type used when indexing; controls the search pipeline.
         """
         self.model: Encoder = model
         self.chunks: list[Chunk] = chunks
@@ -102,10 +114,10 @@ class SembleIndex:
         :param path: Root directory to index.
         :param model: Embedding model to use. Defaults to potion-code-16M.
         :param extensions: File extensions to include. Defaults to a standard set of code extensions.
-        :param content: Content type(s) to index — ``ContentType.CODE`` (default), ``ContentType.DOCS``,
-            ``ContentType.ALL``, or a list of multiple types.
-        :param include_text_files: Deprecated. Use ``content=ContentType.ALL`` instead.
-        :return: An indexed SembleIndex. Chunk file paths are relative to ``path``.
+        :param content: Content type to index: ContentType.CODE (default), ContentType.DOCS,
+            or ContentType.ALL.
+        :param include_text_files: Deprecated. Use content=ContentType.ALL instead.
+        :return: An indexed SembleIndex. Chunk file paths are relative to path.
         :raises FileNotFoundError: If `path` does not exist.
         :raises NotADirectoryError: If `path` exists but is not a directory.
         """
@@ -141,16 +153,16 @@ class SembleIndex:
 
         The repository is cloned into a temporary directory that is removed once
         indexing finishes. Chunk content is preserved in-memory, but
-        ``chunk.file_path`` will not point to a readable file after this call
+        chunk.file_path will not point to a readable file after this call
         returns — it is a repo-relative label, not a filesystem path.
 
         :param url: URL of the git repository to clone (any git provider).
         :param ref: Branch or tag to check out. Defaults to the remote HEAD.
         :param model: Embedding model to use. Defaults to potion-code-16M.
         :param extensions: File extensions to include. Defaults to a standard set of code extensions.
-        :param content: Content type(s) to index — ``ContentType.CODE`` (default), ``ContentType.DOCS``,
-            ``ContentType.ALL``, or a list of multiple types.
-        :param include_text_files: Deprecated. Use ``content=ContentType.ALL`` instead.
+        :param content: Content type to index: ContentType.CODE (default), ContentType.DOCS,
+            or ContentType.ALL.
+        :param include_text_files: Deprecated. Use content=ContentType.ALL instead.
         :return: An indexed SembleIndex. Chunk file paths are repo-relative (e.g. ``src/foo.py``).
         :raises RuntimeError: If git is not on PATH, the clone fails, or times out.
         """
@@ -227,7 +239,7 @@ class SembleIndex:
         :param filter_paths: Optional list of repo-relative file paths; if set, only
             chunks from these files are returned.
         :param rerank: Apply code-tuned reranking (file boost, identifier boost, path penalties).
-            Defaults to ``True`` when ``ContentType.CODE`` was indexed.
+            Defaults to ``True`` when ContentType.CODE was indexed.
         :param diversity: DPP diversity weight in [0, 1]; re-ranks with pyversity after reranking.
             ``None`` (default) auto-detects: uses ``DEFAULT_DOCS_DIVERSITY`` when docs were indexed.
             Pass ``0.0`` to disable diversity even on a docs index.
